[L1] Move public headers on to interface directory
