+++ L1Trigger/DTTrackFinder/src/L1MuDTSectorProcessor.cc
@@ -30,7 +30,7 @@
 //-------------------------------
 
 #include <FWCore/Framework/interface/Event.h>
-#include "L1Trigger/DTTrackFinder/src/L1MuDTTFConfig.h"
+#include "L1Trigger/DTTrackFinder/interface/L1MuDTTFConfig.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTSecProcId.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTSectorReceiver.h"
 #include "L1Trigger/DTTrackFinder/src/L1MuDTDataBuffer.h"
